use a bullmq task instead of polling
